DXP-1568 Load and publish separate descriptions also for product variants
