+++ src/catalog/Model/Attributes/ProductAttributes.php
@@ -6,18 +6,9 @@ use StreamX\ConnectorCatalog\Model\SystemConfig\CatalogConfig;
 
 class ProductAttributes extends BaseProductAttributes
 {
-    const ALWAYS_LOADED_ATTRIBUTES = [
-        'name',
-        'image',
-        'description',
-        'price',
-        'url_key',
-        'media_gallery'
-    ];
-
     public function __construct(CatalogConfig $catalogConfiguration)
     {
-        parent::__construct($catalogConfiguration, self::ALWAYS_LOADED_ATTRIBUTES);
+        parent::__construct($catalogConfiguration);
     }
 
     protected function getConfiguredAttributes(int $storeId): array
