+++ src/catalog/Model/Config/Source/Product/ChildProductAttributeSource.php
@@ -6,24 +6,8 @@ use Magento\Catalog\Api\Data\ProductAttributeInterface;
 use Magento\Catalog\Model\ResourceModel\Product\Attribute\CollectionFactory;
 use StreamX\ConnectorCatalog\Model\ResourceModel\Product as Resource;
 
-class ConfigurableChildProductAttributes extends AbstractAttributeSource
+class ChildProductAttributeSource extends BaseProductAttributeSource
 {
-    const ATTRIBUTES_NOT_ALLOWED_IN_SELECT_LIST = [
-        // always loaded child product attributes - don't allow the user to select them or not
-        'name',
-        'image',
-        'price',
-        'url_key',
-        'media_gallery',
-
-        // explicitly not allowed attributes
-        'gallery',
-        'category_ids',
-        'swatch_image',
-        'quantity_and_stock_status',
-        'options_container',
-    ];
-
     private Resource $productResource;
 
     public function __construct(CollectionFactory $collectionFactory, Resource $productResource)
@@ -37,7 +21,7 @@ class ConfigurableChildProductAttributes extends AbstractAttributeSource
      */
     public function isAllowedInSelectList(ProductAttributeInterface $attribute): bool
     {
-        if (in_array($attribute->getAttributeCode(), self::ATTRIBUTES_NOT_ALLOWED_IN_SELECT_LIST)) {
+        if (!parent::isAllowedInSelectList($attribute)) {
             return false;
         }
