+++ src/catalog/Model/Config/Source/Product/ProductAttributeSource.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace StreamX\ConnectorCatalog\Model\Config\Source\Product;
+
+class ProductAttributeSource extends BaseProductAttributeSource
+{
+}
