+++ src/catalog/etc/adminhtml/system.xml
@@ -32,14 +32,14 @@
                 </field>
                 <field id="product_attributes" translate="label comment" type="multiselect" sortOrder="102" showInDefault="1" showInWebsite="1" showInStore="1">
                     <label>Product attributes to export</label>
-                    <source_model>StreamX\ConnectorCatalog\Model\Config\Source\Product\ConfigurableProductAttributes</source_model>
+                    <source_model>StreamX\ConnectorCatalog\Model\Config\Source\Product\ProductAttributeSource</source_model>
                     <backend_model>StreamX\ConnectorCatalog\Model\Config\Backend\Attributes</backend_model>
                     <can_be_empty>1</can_be_empty>
                     <comment>Select which product attributes should be exported. Some attributes will always be exported: name, image, description, price, url_key, media_gallery</comment>
                 </field>
                 <field id="child_product_attributes" translate="label comment" type="multiselect" sortOrder="103" showInDefault="1" showInWebsite="1" showInStore="1">
                     <label>Configurable children product attributes to export</label>
-                    <source_model>StreamX\ConnectorCatalog\Model\Config\Source\Product\ConfigurableChildProductAttributes</source_model>
+                    <source_model>StreamX\ConnectorCatalog\Model\Config\Source\Product\ChildProductAttributeSource</source_model>
                     <backend_model>StreamX\ConnectorCatalog\Model\Config\Backend\Attributes</backend_model>
                     <can_be_empty>1</can_be_empty>
                     <comment>Select which child (variant) product attributes should be exported. Some attributes will always be exported: name, image, price, url_key, media_gallery and attributes used for creating configurable children (like color, size, etc)</comment>
