+++ test/catalog/integration/utils/MagentoMySqlQueryExecutor.php
@@ -274,6 +274,11 @@ class MagentoMySqlQueryExecutor {
                SET path = '$rootPath/$entityId'
              WHERE entity_id = {$entityIds->getEntityId()}
         ");
+
+        // insert basic attributes
+        self::insertVarcharCategoryAttribute($entityIds->getLinkFieldId(), self::getCategoryAttributeId('display_mode'), 0, 'PRODUCTS');
+        self::insertIntCategoryAttribute($entityIds->getLinkFieldId(), self::getCategoryAttributeId('include_in_menu'), 0, 1);
+
         return $entityIds;
     }
